+++ src/exchange/OrderData.h
@@ -4,13 +4,13 @@
 
 class OrderData {
     public:
-        OrderData(OrderTypes::OrderType type, int price, double volume);
+        OrderData(OrderTypes::OrderType type, long price, double volume);
         ~OrderData();
         const OrderTypes::OrderType getOrderType() const;
-        const int getOrderPrice() const;
+        const long getOrderPrice() const;
         const double getOrderVolume() const;
     private:
         const OrderTypes::OrderType orderType;
-        const int orderPrice;
+        const long orderPrice;
         const double orderVolume;
 };
